third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-end-after-get-set-expected.txt

Test that setting and getting grid-column-end and grid-row-end works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-column-end and grid-row-end set through CSS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-column-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'

Test getting and setting grid-column-end and grid-row-end through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test setting grid-column-start and grid-row-start to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test setting grid-column-start and grid-row-start to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is gridColumnStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is gridColumnEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is gridRowStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is gridRowEndValue

Test setting grid-column-end and grid-row-end back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-column-end') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / 18'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-end') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / 66'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'
PASS successfullyParsed is true

TEST COMPLETE
